nanosseconds

+++ src/DtoResposta.java
@@ -1,6 +1,6 @@
 public record DtoResposta(
         String label,
         Rota rota,
-        long timmer
+        double timmer
 ) {
 }
